Refactoring pages
