+++ web/src/routes/admin/server-status/+page.svelte
@@ -1,7 +1,29 @@
+<script lang="ts">
+	import ServerStatsPanel from '$lib/components/admin-page/server-stats/server-stats-panel.svelte';
+	import { api, ServerStatsResponseDto } from '@api';
+	import { onMount } from 'svelte';
+	import { page } from '$app/stores';
+
+	let serverStat: ServerStatsResponseDto;
+
+	onMount(() => {
+		getServerStats();
+	});
+
+	const getServerStats = async () => {
+		try {
+			const res = await api.serverInfoApi.getStats();
+			serverStat = res.data;
+		} catch (e) {
+			console.log(e);
+		}
+	};
+</script>
+
 <svelte:head>
-  <title>Jobs Status - Immich</title>
+	<title>Jobs Status - Immich</title>
 </svelte:head>
 
-<section>
-  <h1>server status</h1>
-</section>
+{#if $page.data.allUsers && serverStat}
+	<ServerStatsPanel stats={serverStat} allUsers={$page.data.allUsers} />
+{/if}
